edutech/backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
asyncpg==0.31.0
beautifulsoup4==4.14.3
behave==1.3.3
behave-django==1.9.0
boto3==1.35.0
botocore==1.35.99
certifi==2026.2.25
celery
redis
channels
channels-redis
daphne
websockets
cryptography
charset-normalizer==3.4.7
colorama==0.4.6
cucumber-expressions==19.0.0
cucumber-tag-expressions==9.1.0
Django==6.0.3
django-cors-headers==4.4.0
django-environ==0.13.0
django-filter==25.2
django-storages==1.14.4
djangorestframework==3.15.2
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.29
langchain-ollama==1.1.0
langchain-postgres==0.0.17
langchain-text-splitters==1.1.1
langsmith==0.7.31
numpy==2.4.4
ollama==0.6.1
orjson==3.11.8
packaging==26.1
parse==1.21.1
parse_type==0.6.6
pgvector==0.3.6
pillow==12.2.0
pluggy==1.6.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
psycopg2-binary==2.9.11
pydantic==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
PyMuPDF==1.27.2.2
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
s3transfer==0.10.4
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0